add auth & validations for company endpoints

+++ app/routers/company.py
@@ -6,18 +6,34 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from database import get_async_db_context, get_db_context
 from models.company import CompanyModel, CompanyViewModel
-from services.exception import ResourceNotFoundError
+from services.exception import ResourceNotFoundError, AccessDeniedError
 from services import company as CompanyService
+from schemas.user import User
+from services import auth as AuthService
 
 router = APIRouter(prefix="/company", tags=["Company"])
 
 @router.get("", response_model=list[CompanyViewModel])
-async def get_all_company(async_db: AsyncSession = Depends(get_async_db_context)):
+async def get_all_company(
+    async_db: AsyncSession = Depends(get_async_db_context),
+    user: User = Depends(AuthService.token_interceptor)
+    ):
+    
+    if not user.is_admin:
+        raise AccessDeniedError()
+    
     return await CompanyService.get_company(async_db)
 
 
 @router.get("/{company_id}", status_code=status.HTTP_200_OK, response_model=CompanyViewModel)
-async def get_company_by_id(company_id: UUID, db: Session = Depends(get_db_context)):    
+async def get_company_by_id(
+    company_id: UUID,
+    db: Session = Depends(get_db_context),
+    user: User = Depends(AuthService.token_interceptor)): 
+      
+    if not user.is_admin:
+        raise AccessDeniedError()
+    
     company = CompanyService.get_company_by_id(db, company_id)
 
     if company is None:
@@ -27,7 +43,14 @@ async def get_company_by_id(company_id: UUID, db: Session = Depends(get_db_conte
 
 
 @router.post("", status_code=status.HTTP_201_CREATED, response_model=CompanyViewModel)
-async def create_company(request: CompanyModel, db: Session = Depends(get_db_context)):
+async def create_company(
+    request: CompanyModel,
+    db: Session = Depends(get_db_context),
+    user: User = Depends(AuthService.token_interceptor)):
+    
+    if not user.is_admin:
+        raise AccessDeniedError()
+    
     return CompanyService.add_new_company(db, request)
 
 
@@ -36,10 +59,23 @@ async def update_company(
     company_id: UUID,
     request: CompanyModel,
     db: Session = Depends(get_db_context),
+    user: User = Depends(AuthService.token_interceptor)
     ):
-        return CompanyService.update_company(db, company_id, request)
+    
+    if not user.is_admin:
+        raise AccessDeniedError()
+    
+    return CompanyService.update_company(db, company_id, request)
 
 
 @router.delete("/{company_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_company(company_id: UUID, db: Session = Depends(get_db_context)):
+async def delete_company(
+    company_id: UUID,
+    db: Session = Depends(get_db_context),
+    user: User = Depends(AuthService.token_interceptor)
+    ):
+    
+    if not user.is_admin:
+        raise AccessDeniedError()
+    
     CompanyService.delete_company(db, company_id)
